requirements.txt
aiohttp==3.6.2
aiojobs==0.2.2
amqp==2.6.0
appnope==0.1.0
async-timeout==3.0.1
attrs==19.3.0
backcall==0.1.0
billiard==3.6.3.0
cachetools==4.1.0
celery==4.4.4
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
decorator==4.4.2
Flask==1.1.2
flower==0.9.4
future==0.18.2
google-auth==1.16.1
humanize==0.5.1
idna==2.9
ipdb==0.13.2
ipython==7.15.0
ipython-genutils==0.2.0
iso8601==0.1.12
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
kombu==4.6.10
kopf==1.30
kubernetes==11.0.0
MarkupSafe==1.1.1
multidict==4.7.6
oauthlib==3.1.0
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pykube-ng==20.5.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.15.0
tornado==6.0.4
traitlets==4.3.3
typing-extensions==3.7.4.2
urllib3==1.25.9
vine==1.3.0
wcwidth==0.2.3
websocket-client==0.57.0
Werkzeug==1.0.1
yarl==1.4.2
